requirements.txt
flask
Flask-SQLAlchemy
flask-login
Flask-Mail
premailer